show_subscription/requirements.txt
azure-identity
azure-mgmt-resource
